refactor: move listener out of components
