+++ apps/web/ui/menu/context-menu.tsx
@@ -1,6 +1,33 @@
 import { flatten } from '@/utils/tailwind'
 import { Menu } from '@headlessui/react'
-import { ReactElement } from 'react'
+import { showContextMenu } from '@/store/modal'
+import { Dispatch } from 'redux'
+import type { ReactElement } from 'react'
+import type { MouseEvent } from 'react'
+import type { GraphFlowNode } from '@modtree/types'
+import type { ContextMenuType } from 'types'
+
+/**
+ * context:
+ * Context menu and right-click menu are actually the same thing
+ */
+
+export function onContextMenu(
+  dispatch: Dispatch,
+  event: MouseEvent<Element, globalThis.MouseEvent>,
+  menu: ContextMenuType,
+  node?: GraphFlowNode
+) {
+  event.preventDefault()
+  dispatch(
+    showContextMenu({
+      left: event.pageX,
+      top: event.pageY,
+      menu,
+      flowNode: node,
+    })
+  )
+}
 
 export default function ContextMenu(props: {
   children: ReactElement[] | ReactElement
